Fix positive test percentage (#26)
Previously the percentage was calculated between the number of *new*
positive cases and the total number of tests executed on a specific
date. But tests on a specific date could also include people who have
already tested positive, and therefore do not count in the *new*
positive cases. So adjusted to use the total number of people tested,
where second tests for the same person do not count.